requirements.txt
streamlit
google-genai
openai
youtube_transcript_api
python-dotenv
beautifulsoup4
st-copy-to-clipboard
strip-markdown